+++ server/src/main/java/com/linbit/linstor/storage/data/provider/diskless/DisklessData.java
@@ -87,7 +87,7 @@ public class DisklessData<RSC extends AbsResource<RSC>>
             null,
             null,
             discGran.get(),
-            storPool.get().getApiData(null, null, accCtxRef, null, null)
+            storPool.get().getApiData(null, null, accCtxRef, null, null, null, null)
         );
     }
 }