+++ server/src/main/java/com/linbit/linstor/storage/data/provider/exos/ExosData.java
@@ -93,7 +93,7 @@ public class ExosData<RSC extends AbsResource<RSC>>
             getSnapshotUsableSize(),
             new ArrayList<>(getStates()).toString(), // avoid "TransactionList " in the toString()
             discGran.get(),
-            storPool.get().getApiData(null, null, accCtxRef, null, null)
+            storPool.get().getApiData(null, null, accCtxRef, null, null, null, null)
         );
     }
 }