+++ server/src/main/java/com/linbit/linstor/storage/data/provider/file/FileData.java
@@ -77,7 +77,7 @@ public class FileData<RSC extends AbsResource<RSC>>
             getSnapshotUsableSize(),
             new ArrayList<>(getStates()).toString(), // avoid "TransactionList " in the toString()
             discGran.get(),
-            storPool.get().getApiData(null, null, accCtxRef, null, null)
+            storPool.get().getApiData(null, null, accCtxRef, null, null, null, null)
         );
     }
 }