+++ server/src/main/java/com/linbit/linstor/storage/data/provider/spdk/SpdkData.java
@@ -87,7 +87,7 @@ public class SpdkData<RSC extends AbsResource<RSC>>
             getSnapshotUsableSize(),
             new ArrayList<>(getStates()).toString(), // avoid "TransactionList " in the toString()
             discGran.get(),
-            storPool.get().getApiData(null, null, accCtxRef, null, null)
+            storPool.get().getApiData(null, null, accCtxRef, null, null, null, null)
         );
     }