+++ server/src/main/java/com/linbit/linstor/storage/data/provider/storagespaces/StorageSpacesData.java
@@ -77,7 +77,7 @@ public class StorageSpacesData<RSC extends AbsResource<RSC>>
                 getSnapshotAllocatedSize(),
                 getSnapshotUsableSize(),
                 new ArrayList<>(getStates()).toString(), // avoid "TransactionList " in the toString()
-                storPool.get().getApiData(null, null, accCtxRef, null, null)
+                storPool.get().getApiData(null, null, accCtxRef, null, null, null, null)
             );
         }
         if (providerKind.equals(DeviceProviderKind.STORAGE_SPACES_THIN))
@@ -90,7 +90,7 @@ public class StorageSpacesData<RSC extends AbsResource<RSC>>
                 getSnapshotAllocatedSize(),
                 getSnapshotUsableSize(),
                 new ArrayList<>(getStates()).toString(), // avoid "TransactionList " in the toString()
-                storPool.get().getApiData(null, null, accCtxRef, null, null)
+                storPool.get().getApiData(null, null, accCtxRef, null, null, null, null)
             );
         }
         return null;